Loads more refactoring

+++ SteTech1/src/com/scs/stevetech1/components/ICausesHarmOnContact.java
@@ -10,7 +10,7 @@ public interface ICausesHarmOnContact {
 
 	float getDamageCaused();
 	
-	int getSide(); // Prevent friendly-fire
+	byte getSide(); // Prevent friendly-fire
 	
 	IEntity getActualShooter(); // We might be a snowball, but it's the avatar holding the gun that we want
