+++ SteTech1/src/com/scs/stevetech1/components/IDontCollideWithComrades.java
@@ -7,6 +7,6 @@ package com.scs.stevetech1.components;
  */
 public interface IDontCollideWithComrades {
 
-	int getSide();
+	byte getSide();
 	
 }
