+++ SteTech1/src/com/scs/stevetech1/netmessages/GameOverMessage.java
@@ -10,14 +10,14 @@ import com.jme3.network.serializing.Serializable;
 @Serializable
 public class GameOverMessage extends MyAbstractMessage {
 	
-	public int winningSide;
+	public byte winningSide;
 	
 	public GameOverMessage() {
 		super(true, true);
 	}
 
 
-	public GameOverMessage(int _winningSide) {
+	public GameOverMessage(byte _winningSide) {
 		super(true, true);
 		
 		winningSide = _winningSide;
