+++ SteTech1/src/com/scs/stevetech1/server/ClientData.java
@@ -39,7 +39,7 @@ public class ClientData {
 	}
 	
 	
-	public int getSide() {
+	public byte getSide() {
 		return this.playerData.side;
 	}
