+++ TestGame/src/com/scs/testgame/entities/RoamingZombie.java
@@ -104,7 +104,7 @@ public class RoamingZombie extends PhysicalEntity implements IAffectedByPhysics,
 
 
 	@Override
-	public int getSide() {
+	public byte getSide() {
 		return 0;
 	}
