+++ TestGame/src/com/scs/testgame/entities/TestGameEnemyAvatar.java
@@ -10,7 +10,7 @@ public class TestGameEnemyAvatar extends AbstractOtherPlayersAvatar {
 	//private ChronologicalLookup<HistoricalAnimationData> animList = new ChronologicalLookup<HistoricalAnimationData>(true, -1);
 	private CharacterModel model;
 
-	public TestGameEnemyAvatar(IEntityController game, int eid, float x, float y, float z, int side) {
+	public TestGameEnemyAvatar(IEntityController game, int eid, float x, float y, float z, byte side) {
 		super(game, TestGameClientEntityCreator.AVATAR, eid, x, y, z, new CharacterModel(game.getAssetManager()), side, "TestGame");
 		
 		model = (CharacterModel)super.anim;
