+++ Tutorial/src/boxwars/entities/BoxWarsClientAvatar.java
@@ -11,7 +11,7 @@ import boxwars.models.BoxAvatarModel;
 
 public class BoxWarsClientAvatar extends AbstractClientAvatar {
 
-	public BoxWarsClientAvatar(AbstractGameClient _module, int _playerID, IInputDevice _input, Camera _cam, int eid, float x, float y, float z, int side) {
+	public BoxWarsClientAvatar(AbstractGameClient _module, int _playerID, IInputDevice _input, Camera _cam, int eid, float x, float y, float z, byte side) {
 		super(_module, BoxWarsServer.AVATAR, _playerID, _input, _cam, eid, x, y, z, side, new BoxAvatarModel(_module.getAssetManager()), new PersonAvatar(_module, _input, 3f, 2f));
 		
 	}
