+++ Tutorial/src/boxwars/entities/BoxWarsServerAvatar.java
@@ -15,7 +15,7 @@ public class BoxWarsServerAvatar extends AbstractServerAvatar {
 	public BoxWarsServerAvatar(BoxWarsServer _module, ClientData _client, IInputDevice _input, int eid) {
 		super(_module, BoxWarsServer.AVATAR, _client, _input, eid, new BoxAvatarModel(_module.getAssetManager()), 1f, 1, new PersonAvatar(_module, _input, 3f, 2f));
 
-		IAbility abilityGun = new PlayersGun(_module, _module.getNextEntityID(), client.getPlayerID(), this, eid, 0, client);
+		IAbility abilityGun = new PlayersGun(_module, _module.getNextEntityID(), client.getPlayerID(), this, eid, (byte)0, client);
 		_module.actuallyAddEntity(abilityGun);
 		
 	}
