+++ UndercoverAgent/src/com/scs/undercoveragent/entities/SnowmanClientAvatar.java
@@ -11,7 +11,7 @@ import com.scs.undercoveragent.models.SnowmanModel;
 
 public class SnowmanClientAvatar extends AbstractClientAvatar {
 	
-	public SnowmanClientAvatar(AbstractGameClient _module, int _playerID, IInputDevice _input, Camera _cam, int eid, float x, float y, float z, int side) {
+	public SnowmanClientAvatar(AbstractGameClient _module, int _playerID, IInputDevice _input, Camera _cam, int eid, float x, float y, float z, byte side) {
 		super(_module, UndercoverAgentClientEntityCreator.AVATAR, _playerID, _input, _cam, eid, x, y, z, side, new SnowmanModel(_module.getAssetManager()), new PersonAvatar(_module, _input, UAStaticData.MOVE_SPEED, UAStaticData.JUMP_FORCE));
 		
 	}
